:green_heart: chore: ContactChannel 섹션 폴더 분리 #26
